Replace Int64 with Long

+++ here-naksha-handler-activitylog/src/jvmMain/java/com/here/naksha/handler/activitylog/ActivityLogComparator.java
@@ -38,7 +38,7 @@ public class ActivityLogComparator implements Comparator<NakshaFeature> {
   }
 
   private static long updatedAt(NakshaFeature feature) {
-    return xyzNamespace(feature).getUpdatedAt().toLong();
+    return xyzNamespace(feature).getUpdatedAt();
   }
 
   private static XyzNs xyzNamespace(NakshaFeature feature) {
